feat: eHandle.screenshot({fullElement: true})
Implements a 'fullElement' option to ElementHandle::screenshot (but not
Page::screenshot). This works similarly to `fullPage` in
Page::screenshot,
except instead of setting device metrics to the page content size,
it sets them to the size of an element. If the element is smaller than
the current viewport, we don't make the viewport smaller.